requirements.txt
Flask>=2.0.1
pytesseract>=0.3.8
Pillow>=10.0.0
pandas>=1.5.3
numpy>=1.24.0
langchain-openai>=0.0.2
python-dotenv>=0.19.0
streamlit>=1.28.0